refactor: enums 디렉토리 -> entity 디렉토리 하위로 이동
