+++ src/main/java/com/soongsil/CoffeeChat/dto/MenteeRequest.java
@@ -1,6 +1,6 @@
 package com.soongsil.CoffeeChat.dto;
 
-import com.soongsil.CoffeeChat.enums.PartEnum;
+import com.soongsil.CoffeeChat.entity.enums.PartEnum;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
