+++ src/main/java/com/soongsil/CoffeeChat/dto/UserRequest.java
@@ -1,7 +1,7 @@
 package com.soongsil.CoffeeChat.dto;
 
-import com.soongsil.CoffeeChat.enums.ClubEnum;
-import com.soongsil.CoffeeChat.enums.PartEnum;
+import com.soongsil.CoffeeChat.entity.enums.ClubEnum;
+import com.soongsil.CoffeeChat.entity.enums.PartEnum;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
