+++ src/main/java/com/soongsil/CoffeeChat/entity/enums/ClubEnum.java
@@ -1,4 +1,4 @@
-package com.soongsil.CoffeeChat.enums;
+package com.soongsil.CoffeeChat.entity.enums;
 
 public enum ClubEnum {
     GDSC,
